backend/requirements.txt
Flask==3.1.0
flask-cors==6.0.0
PyJWT==2.9.0
pymongo==3.11.4
bcrypt==4.3.0
python-dotenv==1.0.0
pydantic